(refactor) Added new folders to the workspace
