+++ src/app/app.routes.ts
@@ -3,6 +3,6 @@ import { Routes } from '@angular/router';
 export const routes: Routes = [
   {
     path: '',
-    loadChildren: () => import('./tabs/tabs.routes').then((m) => m.routes)
+    loadChildren: () => import('./pages/tabs/tabs.routes').then((m) => m.routes)
   }
 ];
